update PSR, moved destroy method to ApiDonationController
