+++ routes/api.php
@@ -2,9 +2,6 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Donation\DonationTableController;
-use App\Http\Controllers\DonationController;
-use App\Services\DonationService;
 use App\Http\Controllers\Api\DonationController as ApiDonationController;
 
 /*
@@ -22,6 +19,6 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::delete('/donations/{id}', [DonationController::class, 'destroy']);
+Route::delete('/donations/{id}', [ApiDonationController::class, 'destroy']);
 Route::get('/donations', [ApiDonationController::class, 'index']);
 Route::get('/donations/widget', [ApiDonationController::class, 'getWidgetData']);
